Update BubbleSort.java
Commented out, color function calls and swapped them for, system.out.println statement, Katie check Console for the __dump(array) function, I'm not sure if you want to completely eliminate console if that one function is necessary, I'm not too sure what it does to be honest.  But if you wanted to change this to having one main and that is necessary you could just keep that one piece of functionality.