Fix Issue #14 ; Partial Fix Issue #2
This commit fixes Issue #14 but does not completely fix #2. For #2, the
Menu.java file has been created and a single main function can execute
the entire program; however, not all of the files are being called as I
am waiting for #4 to be resolved. The fact that these functions are not
being called shouldn't break the program; however, nothing will happen
if you attempt to run some of the Data Structures functions.