Semi-Fix #18
Replaced the purposeful error throw when a valid value is not received.
As long as the user continues to enter an integer value outside the
range of the ArrayList, they will continue to be prompted for an
integer. An error is thrown if they attempt to enter a string, which
sends them back to the main menu. Although the string error is not
"fixed," the user is warned that it will happen if they do not enter an
integer. This warning was also added to BinaryTree.java.